refactor: modularize scripts and enable modules
