+++ eslint.config.js
@@ -4,10 +4,11 @@ export default [
     languageOptions: {
       ecmaVersion: "latest",
       sourceType: "module",
+      globals: { window: "readonly", document: "readonly" },
     },
     rules: {
       semi: ["error", "always"],
-      quotes: ["error", "double"],
+      quotes: "off",
     },
   },
 ];
