+++ index.html
@@ -635,6 +635,6 @@
     <!-- Markdown Parser Library (for AI Markdown output) -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/marked/12.0.1/marked.min.js" crossorigin="anonymous"></script>
     <!-- JavaScript File -->
-    <script src="js/main.js"></script>
+    <script type="module" src="js/components/main.js"></script>
 </body>
 </html>
